# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.Alert:
    properties:
      acknowledged_at:
        type: string
      acknowledged_by:
        type: string
      id:
        type: integer
      is_acknowledged:
        type: boolean
      locomotive_id:
        type: integer
      message:
        type: string
      parameter_name:
        type: string
      resolved_at:
        type: string
      severity:
        description: warning | critical
        type: string
      threshold_value:
        type: number
      ts:
        type: string
      value:
        type: number
    type: object
  domain.AuthResponse:
    properties:
      token:
        type: string
    type: object
  domain.Event:
    properties:
      created_by:
        type: string
      description:
        type: string
      event_type:
        description: incident | maintenance | note | replay_mark
        type: string
      id:
        type: integer
      locomotive_id:
        type: integer
      telemetry_snapshot:
        type: object
      ts:
        type: string
    type: object
...
  domain.Locomotive:
    properties:
      depot:
        type: string
      id:
        type: integer
      manufacturer:
        type: string
      model:
        type: string
      number:
        type: string
      power_type:
        description: diesel | electric
        type: string
      status:
        description: active | maintenance
        type: string
      type:
        description: cargo | passenger
        type: string
      year_built:
        type: integer
    type: object
  domain.LoginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  domain.RegisterRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Real-time telemetry dashboard for locomotive dispatchers and drivers.
  title: Locomotive Digital Twin API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token
      parameters:
      - description: Login and password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account and returns a JWT token
      parameters:
      - description: Login and password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Login already taken
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
...
  /locomotives:
    get:
      description: Returns all locomotives with their current status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Locomotive'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all locomotives
      tags:
      - locomotives
  /locomotives/{id}:
    get:
      description: Returns a single locomotive by its ID
      parameters:
      - description: Locomotive ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Locomotive'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get locomotive by ID
      tags:
      - locomotives
  /locomotives/{id}/alerts:
    get:
      description: Returns alerts for a locomotive. Filter by active status and/or
        severity.
      parameters:
      - description: Locomotive ID
        in: path
        name: id
        required: true
        type: integer
      - description: Only unacknowledged alerts
        in: query
        name: active
        type: boolean
      - description: 'Filter by severity: warning or critical'
        in: query
        name: severity
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Alert'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List alerts
      tags:
      - alerts
  /locomotives/{id}/alerts/{alert_id}/acknowledge:
    post:
      consumes:
      - application/json
      description: Marks an alert as acknowledged by the dispatcher
      parameters:
      - description: Locomotive ID
        in: path
        name: id
        required: true
        type: integer
      - description: Alert ID
        in: path
        name: alert_id
        required: true
        type: integer
      - description: Acknowledger name
        in: body
        name: body
        schema:
          properties:
            acknowledged_by:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Acknowledge an alert
      tags:
      - alerts
  /locomotives/{id}/events:
    get:
      description: Returns manual events (incidents, maintenance notes) for a locomotive
      parameters:
      - description: Locomotive ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Start time (RFC3339), default: 1h ago'
        in: query
        name: from
        type: string
      - description: 'End time (RFC3339), default: now'
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Event'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List events
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Records a manual event (incident, maintenance, note) for a locomotive
      parameters:
      - description: Locomotive ID
        in: path
        name: id
        required: true
        type: integer
      - description: Event payload
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/domain.Event'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Event'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create manual event
      tags:
      - events
  /locomotives/{id}/export/csv:
    get:
      description: Downloads a CSV file containing historical telemetry for a locomotive
      parameters:
      - description: Locomotive ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Start time (RFC3339), default: 1h ago'
        in: query
        name: from
        type: string
      - description: 'End time (RFC3339), default: now'
        in: query
        name: to
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Export telemetry as CSV
      tags:
      - export
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"